Let there be defuns
Our Emacs, who art in RAM,
Hallowed be thy commands.
Thy buffers come,
Thy shortcuts be done,
In code as it is in text.

Give us this day our daily M-x.
And forgive us our typos,
As we forgive those who Vim against us.
And lead us not into evil modes,
But deliver us from Emacs' demise.

For thine is the keyboard,
And the lisp,
And the meta key,
Forever and ever.

Ctrl-x Ctrl-c